Never use string literal building in substitutions
The purpose of this commit was initially to fix a bug where the "'"
character would break links.  For example, trying to make the link
[[This doesn't work]] by pressing <CR> on a visual selection of those
words would cause a syntax error.  After fixing this, I found many more
instances of the problematic substitution pattern that attempted to
build a string literal for use with \=.  I grepped around and I think I
have fixed all of them.  Any bugs this change cause will probably be
minimal, but it is possible I missed a string escaping mechanism
somewhere like one I found that replaced "'" with "''" as a workaround
solution for one example of the bug this commit attempts to fix.