Fix linking in visual line mode
Turns out there really was a reason why linking was only defined for
characterwise-visual mode.  This fixes the s:normalize_link_syntax_v
functions so that they know what visual mode they were in last and gets
rid of the pesky trailing newline that you get from linewise-visual
mode.  Removing newlines doesn't matter because except for visual mode
it should be impossible to have a newline in the selection.