Fix cursor positioning in vimwiki#lst#change_level
In the function vimwiki#lst#change_level the cur_col was using the
column the cursor was in (`col('.')`). This is a problem when moving
from insert mode to normal mode however when the last column no longer
exists, this means that changing the level from the last column should
change the cursor position. This result is apparent in Neovim, but
doesn't occur in Vim.

By calculating cur_col from the last insert position this issue is
avoided in both vim and neovim.