Prevent sticky type checking (E706) errors for old versions of vim
Vim used to throw the E706 error if you tried to change a variable's
type. This error can be triggered by vimwiki if, during the user
settings loop in `read_global_settings_from_user()`, the variable type
for `users_value` changes.

The vim error was removed in v7.4.1546 as part of
https://github.com/vim/vim/commit/f6f32c38bf3319144a84a01a154c8c91939e7acf.