append extension to links containing "." (fixes #950)
If a link contains `.`, then this:

  fnamemodify(link_text, ':e')

will return the last .-separated part of the name as extension, and nothing
else will get appended.  Instead, we check if the extension is empty _or_
doesn't match the extension for the target wiki, in which case we append
the correct one.

This seems like it works, and might not break anything else, given the
branch of the function that it's in.