resolve_link(): check ext against version with dot (fixes #1256)
It turns out that the target extension here in `ext` has a dot prepended.

Under manual testing, this seems to fix the issue where a wiki with
`markdown_link_ext` set and using `.md` (for example) in links will wind up
taking people to pages with a repeated extension, like `foo.md.md`.

I feel like the repeated use of:

    vimwiki#vars#get_wikilocal('ext', link_infos.index)

here should probably be replaced with a named variable like `target_ext`,
but I wanted to touch the minimum lines of code possible for this fix.