Fix error when opening links using system handler on Windows
Under Windows and Vim 9, opening a link that triggers the system handler
fails with a "command not found" error message. It looks like the title
argument to `start` is treated as the executable path.

According to the help page for `start`, the title argument is optional,
and /B instructs start not to open a CMD window, so the title is never
used and can be removed with no loss in functionality.

This change removes the redundant title argument, fixing the "command
not found" error.