Make sure all folds are open before toggling
If folding is enabled when the script jumps to the todo in the original
file, folding can interfere with what is toggled and which lines are
used to update the state in the todo list window/buffer.

Added zR to make sure all and any folds are open before doing anything
else in the original file, and thus making sure only the appropriate
line will be copied back to the quickfix buffer.